Fix int in models

+++ TWITOFF/models.py
@@ -7,7 +7,7 @@ DB = SQLAlchemy()
 
 class User(DB.Model):
     """Twitter users to analyze"""
-    id = DB.Column(DB.Integer, primary_key=True)
+    id = DB.Column(DB.BigInteger, primary_key=True)
     name = DB.Column(DB.String(15), nullable=False)
     newest_tweet_id = DB.Column(DB.BigInteger)
 
@@ -19,7 +19,7 @@ class Tweet(DB.Model):
     """The user's tweets from twitter"""
     id = DB.Column(DB.BigInteger, primary_key=True)
     text = DB.Column(DB.Unicode(300))
-    user_id = DB.Column(DB.Integer, DB.ForeignKey("user.id"), nullable=False)
+    user_id = DB.Column(DB.BigInteger, DB.ForeignKey("user.id"), nullable=False)
     user = DB.relationship("User", backref=DB.backref("tweets", lazy=True))
 
     embedding = DB.Column(DB.PickleType, nullable=False)
